requirements.txt
sqlalchemy==1.0.9
nose
unittest2==0.5.1
psycopg2
factory-boy==2.9.2
